requirements.txt
matplotlib==1.4.3
nltk==3.1
numpy==1.9.2
scikit-learn==0.16.1
scipy==0.16.0b2
tqdm==3.7.1
facebook-sdk==0.4.0
requests==2.7.0